requirements.txt
praw==7.8.1
pyyaml==6.0.1
pandas==2.1.0
supabase==1.0.3
# supabase-py==2.0.3
python-dotenv==1.0.0
httpx>=0.23.1
gotrue==1.1.0 